+++ packages/docs/docs/main/change-log.md
@@ -11,6 +11,8 @@ This page lists the changes to the IsaacScript framework.
 
 ## September 11th, 2023 (Unreleased)
 
+- Added the following helper functions:
+  - `useCardTemp`
 - Renamed the following helper functions:
   - `addStat` --> `addPlayerStat`